Substructure Search with RDKit
Required: RDKit library installed.

The system variable:
One system environment variable PG_CART_INSTALLED is added with the
default value as false. If RDKit is installed and you want to turn on
this feature for user, remember to set RDKIT_INSTALLED as true.

The migration files:
1. They are executed only if RDKit is installed in Postgresql.
2. A new column is added to the table: samples. The script for
generating the data into the new column is included.