Move ToggleButton to tab body and replace it by proper bootstrap component
The original issue was just that the ToggleButton was situated in the Tab Button, which throws an error, because you are not allowed
to nest a <button> within a <button>.

Then it occurred to me, that this was the only instance where this button is used and that it's usage (namely toggling reaction scheme between default and gaseous)
was not very clear until you hover over the button and read the tooltip.

Therefore I replaced it by a regular bootstrap buttongroup we also use in other places. This way, the toggling functionality becomes clear directly,
as both buttons are displayed at the same time. So no more hover necessary. Additionally, the buttons use a global utility component for easier styling
of toggle buttons throughout the app.